build: move sources into subfolder
Since the application is also integrated in the LuCI. But the project
has a different folder structure it is not so easy to port the changes
with patches. To make this work better and to make it easier to apply
the changes in both projects, the folder structure was adapted to the
LuCI project.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>